requirements.txt
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
requests>=2.32
jinja2>=3.1
